test: run slow tests only on demand
This change will skip over the slower tests (e.g. benchmarks) by
default.
On the CI all tests are executed and run in band, while locally all fast
tests are executed in parallel.
