+++ .github/workflows/ci.yml
@@ -35,4 +35,6 @@ jobs:
 
       - run: npm ci
       - run: npm run check
-      - run: npm test
+      - run: npm test -- --run-in-band
+        env:
+          RUN_SLOW_TESTS: 1
